requirements.txt
# Data processing
pandas>=2.0.0
numpy>=1.24.0

# Machine Learning
scikit-learn>=1.3.0
xgboost>=2.0.0


# API
fastapi>=0.100.0
uvicorn>=0.23.0
pydantic>=2.0.0

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0

# Notebooks
jupyter>=1.0.0
matplotlib>=3.7.0
seaborn>=0.12.0

# Utilities
python-dotenv>=1.0.0
pyyaml>=6.0.0
joblib>=1.3.0

